Make junction restrictions structures nested types of JunctionRestrictionsManager (some of these will be moved to new locations later)
