+++ TLM/TLM/TLM.csproj
@@ -147,14 +147,11 @@
     <Compile Include="Manager\Impl\ExtSegmentEndManager.cs" />
     <Compile Include="Manager\Impl\ExtSegmentManager.cs" />
     <Compile Include="Manager\Impl\GeometryNotifier.cs" />
-    <Compile Include="Manager\Impl\JunctionRestrictionFlags.cs" />
     <Compile Include="Manager\Impl\LaneConnection\ConnectionDataBase.cs" />
     <Compile Include="Manager\Impl\LaneConnection\LaneConnectionData.cs" />
     <Compile Include="Manager\Impl\LaneConnection\LaneConnectionSubManager.cs" />
     <Compile Include="Manager\Impl\LaneConnection\LaneConnectionManager.cs" />
     <Compile Include="Manager\Impl\LaneConnection\LaneEnd.cs" />
-    <Compile Include="Manager\Impl\JunctionRestrictions.cs" />
-    <Compile Include="Manager\Impl\SegmentJunctionRestrictions.cs" />
     <Compile Include="Patch\_External\RTramAIModPatch.cs" />
     <Compile Include="State\OptionsTabs\GameplayTab_AIGroups.cs" />
     <Compile Include="State\OptionsTabs\GeneralTab_CompatibilityGroup.cs" />
